fix type in hovering reward

+++ 04_Repository/Gym/Wrappers/hovering_wrapper.py
@@ -22,7 +22,7 @@ class HoveringWrapper(gym.Wrapper):
         obs, info = self.env.reset(seed, options, degrees, position)
         self.hover_count = 0
         # Return the augmented observation
-        augmented_obs = np.append(obs, self.get_hover_ratio())
+        augmented_obs = np.append(obs, self.get_hover_ratio()).astype(np.float32)
         return augmented_obs, info
     
     def get_hover_ratio(self):
@@ -43,5 +43,5 @@ class HoveringWrapper(gym.Wrapper):
         # print(self.hover_count, np.linalg.norm(action))
         # Augment the observation with the hover count
         
-        augmented_obs = np.append(obs, self.get_hover_ratio())
+        augmented_obs = np.append(obs, self.get_hover_ratio()).astype(np.float32)
         return augmented_obs, reward, done, truncated, info
